Move _create_networks_to_routers to OvnClientMixin
Extract code that will be shared by OvnScenario and the new Context.
