+++ rally_ovs/plugins/ovs/ovnclients.py
@@ -15,14 +15,16 @@
 import netaddr
 
 from rally.common import logging
-from rally.common import utils
+from rally.common.utils import RandomNameGeneratorMixin
+
 from rally_ovs.plugins.ovs import ovsclients
+from rally_ovs.plugins.ovs import utils
 
 
 LOG = logging.getLogger(__name__)
 
 
-class OvnClientMixin(ovsclients.ClientsMixin, utils.RandomNameGeneratorMixin):
+class OvnClientMixin(ovsclients.ClientsMixin, RandomNameGeneratorMixin):
 
     def _create_lswitches(self, lswitch_create_args, num_switches=-1):
         self.RESOURCE_NAME_FORMAT = "lswitch_XXXXXX_XXXXXX"
@@ -89,3 +91,42 @@ class OvnClientMixin(ovsclients.ClientsMixin, utils.RandomNameGeneratorMixin):
         ovn_nbctl.enable_batch_mode(False)
 
         return lrouters
+
+    def _connect_network_to_router(self, router, network):
+        LOG.info("Connect network %s to router %s" % (network["name"], router["name"]))
+
+        ovn_nbctl = self.controller_client("ovn-nbctl")
+        install_method = self.install_method
+        ovn_nbctl.set_sandbox("controller-sandbox", install_method)
+        ovn_nbctl.enable_batch_mode(False)
+
+
+        base_mac = [i[:2] for i in self.task["uuid"].split('-')]
+        base_mac[0] = str(hex(int(base_mac[0], 16) & 254))
+        base_mac[3:] = ['00']*3
+        mac = utils.get_random_mac(base_mac)
+
+        lrouter_port = ovn_nbctl.lrouter_port_add(router["name"], network["name"], mac,
+                                                  str(network["cidr"]))
+        ovn_nbctl.flush()
+
+
+        switch_router_port = "rp-" + network["name"]
+        lport = ovn_nbctl.lswitch_port_add(network["name"], switch_router_port)
+        ovn_nbctl.db_set('Logical_Switch_Port', switch_router_port,
+                         ('options', {"router-port":network["name"]}),
+                         ('type', 'router'),
+                         ('address', 'router'))
+        ovn_nbctl.flush()
+
+    def _connect_networks_to_routers(self, lnetworks, lrouters, networks_per_router):
+        j = 0
+        for i in range(len(lrouters)):
+            lrouter = lrouters[i]
+            LOG.info("Connect %s networks to router %s" % (networks_per_router, lrouter["name"]))
+            for k in range(j, j+int(networks_per_router)):
+                lnetwork = lnetworks[k]
+                LOG.info("connect networks %s cidr %s" % (lnetwork["name"], lnetwork["cidr"]))
+                self._connect_network_to_router(lrouter, lnetwork)
+
+            j += int(networks_per_router)
