+++ rally_ovs/plugins/ovs/scenarios/ovn.py
@@ -194,45 +194,10 @@ class OvnScenario(ovnclients.OvnClientMixin, scenario.OvsScenario):
         LOG.info("Create Logical routers")
         return super(OvnScenario, self)._create_routers(router_create_args)
 
-
-    def _connect_network_to_router(self, router, network):
-        LOG.info("Connect network %s to router %s" % (network["name"], router["name"]))
-
-        ovn_nbctl = self.controller_client("ovn-nbctl")
-        install_method = self.install_method
-        ovn_nbctl.set_sandbox("controller-sandbox", install_method)
-        ovn_nbctl.enable_batch_mode(False)
-
-
-        base_mac = [i[:2] for i in self.task["uuid"].split('-')]
-        base_mac[0] = str(hex(int(base_mac[0], 16) & 254))
-        base_mac[3:] = ['00']*3
-        mac = utils.get_random_mac(base_mac)
-
-        lrouter_port = ovn_nbctl.lrouter_port_add(router["name"], network["name"], mac,
-                                                  str(network["cidr"]))
-        ovn_nbctl.flush()
-
-
-        switch_router_port = "rp-" + network["name"]
-        lport = ovn_nbctl.lswitch_port_add(network["name"], switch_router_port)
-        ovn_nbctl.db_set('Logical_Switch_Port', switch_router_port,
-                         ('options', {"router-port":network["name"]}),
-                         ('type', 'router'),
-                         ('address', 'router'))
-        ovn_nbctl.flush()
-
     def _connect_networks_to_routers(self, lnetworks, lrouters, networks_per_router):
-        j = 0
-        for i in range(len(lrouters)):
-            lrouter = lrouters[i]
-            LOG.info("Connect %s networks to router %s" % (networks_per_router, lrouter["name"]))
-            for k in range(j, j+int(networks_per_router)):
-                lnetwork = lnetworks[k]
-                LOG.info("connect networks %s cidr %s" % (lnetwork["name"], lnetwork["cidr"]))
-                self._connect_network_to_router(lrouter, lnetwork)
-
-            j += int(networks_per_router)
+        super(OvnScenario, self)._connect_networks_to_routers(lnetworks,
+                                                              lrouters,
+                                                              networks_per_router)
 
     @atomic.action_timer("ovn_network.create_phynet")
     def _create_phynet(self, lswitches, physnet, batch):
